+++ packages/protocolbeat/src/store/utils/stringHash.ts
