+++ packages/protocolbeat/src/view/NodeView.tsx
@@ -4,7 +4,7 @@ import { useCallback, useRef } from 'react'
 import type { Field, Node } from '../store/State'
 import { useStore } from '../store/store'
 import { NODE_WIDTH } from '../store/utils/constants'
-import { OklchColor, oklchColorToCSS } from '../utils/color'
+import { OklchColor, oklchColorToCSS } from '../store/utils/color'
 import { ResizeHandle } from './ResizeHandle'
 
 export interface NodeViewProps {
